More -Wshorten-64-to-32 fixes on 32 bit builds
git-svn-id: https://svn.osgeo.org/gdal/trunk@31650 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/netcdf/netcdfdataset.cpp
@@ -442,7 +442,7 @@ netCDFRasterBand::netCDFRasterBand( netCDFDataset *poNCDFDS,
         if( ( status == NC_NOERR ) && ( nTmpFormat == NC_CHUNKED ) ) {
             CPLDebug( "GDAL_netCDF", 
                       "setting block size to chunk size : %ld x %ld\n",
-                      chunksize[nZDim-1], chunksize[nZDim-2]);
+                      static_cast<long>(chunksize[nZDim-1]), static_cast<long>(chunksize[nZDim-2]));
             nBlockXSize = (int) chunksize[nZDim-1];
             nBlockYSize = (int) chunksize[nZDim-2];
         }
